feat(board): status-description tooltip in board column headers
Add an optional `tooltip` to BoardColumnDef, rendered as an info-circle
icon button in the expanded column header (before the collapse arrow, per
the tickets Figma). Hovering — or tapping, on touch — shows the status
description in a tooltip.

The touch-toggle tooltip logic that ticket-status-config-row carried
locally is extracted into a shared ui/TouchFriendlyTooltip (with a
configurable side) and reused by both call sites.
