+++ openframe-frontend-core/src/components/features/board/board-column-header.tsx
@@ -2,7 +2,8 @@
 
 import { Button } from '../../ui/button/button'
 import { TicketStatusTag, resolveTicketStatus } from '../../ui/ticket-status-tag'
-import { Arrow02LeftIcon, Arrow02RightIcon, BoxArchiveIcon, PlusIcon } from '../../icons-v2-generated'
+import { TouchFriendlyTooltip } from '../../ui/touch-friendly-tooltip'
+import { Arrow02LeftIcon, Arrow02RightIcon, BoxArchiveIcon, InfoCircleIcon, PlusIcon } from '../../icons-v2-generated'
 import type { BoardColumnDef } from './types'
 
 export interface BoardColumnHeaderProps {
@@ -60,6 +61,18 @@ export function BoardColumnHeader({
         </span>
       </div>
       <div className="flex shrink-0 items-center gap-[var(--spacing-system-xxs)]">
+        {column.tooltip && (
+          <TouchFriendlyTooltip content={column.tooltip} side="bottom">
+            <Button
+              variant="transparent"
+              size="icon"
+              className="h-8 w-8 md:h-8 md:w-8 p-0"
+              aria-label={column.tooltip}
+            >
+              <InfoCircleIcon className="h-6 w-6 text-ods-text-secondary" />
+            </Button>
+          </TouchFriendlyTooltip>
+        )}
         <Button
           variant="transparent"
           size="icon"
