+++ openframe-frontend-core/src/components/ui/index.ts
@@ -116,6 +116,7 @@ export * from './tag'
 export * from './title-content-block'
 export * from './toggle'
 export * from './tooltip'
+export * from './touch-friendly-tooltip'
 export * from './error-state'
 export * from './content-loader'
